# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Provisum API
  version: 0.7.0
  description: |
    REST API for Provisum — enterprise security role mapping platform for ERP migrations
    (e.g. SAP ECC to S/4HANA). Manages the full workflow: data upload, AI persona generation,
    role mapping, SOD conflict analysis, approvals, and provisioning exports.

    **Authentication:** All endpoints (except `/api/auth/login`, `/api/auth/setup`, and
    `/api/health`) require a valid Supabase JWT session cookie. The cookie is set automatically
    on successful login via `/api/auth/login`.

    **Role hierarchy (higher = more access):**
    `system_admin (100) > admin (80) > approver (60) > coordinator (50) > mapper (40) > viewer (20)`
  contact:
    name: Provisum Support
    url: https://provisum.io
  license:
    name: Proprietary
servers:
  - url: https://demo.provisum.io
    description: Demo instance
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: Supabase JWT session cookie (set by /api/auth/login)
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "Unauthorized"
...
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Unauthorized"
...
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
security:
  - cookieAuth: []
paths:
...
  /api/calibration:
    get:
      tags: [Dashboard & Health]
      summary: Get low-confidence persona assignments for calibration
      description: |
        Returns persona assignments below the specified confidence threshold for manual
        review and calibration. Scoped to the current user's org unit.
      parameters:
        - name: threshold
          in: query
          schema:
            type: integer
            default: 70
          description: Confidence score threshold (0-100). Returns assignments at or below this value.
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
          description: Maximum number of results to return
      responses:
        "200":
          description: Low-confidence assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    assignmentId:
                      type: integer
                    userId:
                      type: integer
                    personaId:
                      type: integer
                    confidenceScore:
                      type: number
                    aiReasoning:
                      type: string
                      nullable: true
                    userName:
                      type: string
                    department:
                      type: string
                      nullable: true
                    personaName:
                      type: string
                    personaDescription:
                      type: string
                      nullable: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"

  # ===========================================================================
  # Cron
  # ===========================================================================
  /api/cron/exports:
    get:
      tags: [Admin]
      summary: Process due scheduled exports (cron endpoint)
      description: |
        Called by Vercel cron jobs to process any scheduled exports that are due.
        Authenticated via `CRON_SECRET` bearer token in the Authorization header.
      security:
        - {}
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
          description: "Bearer {CRON_SECRET}"
      responses:
        "200":
          description: Cron run results
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  processed:
                    type: integer
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        name:
                          type: string
                        status:
                          type: string
        "401":
          description: Invalid cron secret
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ===========================================================================
  # Docs (this spec)
  # ===========================================================================
...